Reduce max size of collection that stores the accessions when NO_LIMIT option is chosen.
The NO_LIMIT value is currently Integer.MAX_VALUE if the client doesn't want a limit, then this is the option he will choose. However this value dictates the size of the collection that holds the accession. We can not initialize this to Integer.MAX_VALUE, because its to big. The collection is now reduced to the hit count size if this is smaller than the limit.