Fix executor service shutdown, fix bug on number of threads being started up in pool.
The executor service was not being shutdown properly. The observable method being called to shutdown the Executor service was the onCompleted. However given the immutable nature of Observables the completion of certain queries was never verified. To guarantee shutdown, we now apply this to the onUnsubscribe() of the Observable.

The second issue to be fixed was the calcualtaion of the number of threads the thread pool should initialize with. Before the correction, this value was based on the max value between the default chunk size or the number of query requests.  If the query had a large number of requests the pool size would be very big. I have now changed the value to be the minimum between chunk size and request size.